+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 2.9.21
 
 * Custom items could be duplicated during import.
+* Effect: Storm Sphere effect.
 
 # 2.9.20